+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ### Added
 
-- added some `Quantifier` instance methods
-  - `#options`, `#ts`, `#type`
+- added some `Expression::Base` methods to `Quantifier` instances
+  - `#type`, `#options`, `#terminal?`
+  - `#base_length`, `#full_length`, `#starts_at`, `#ts`, `#offset`
   - `#conditional_level`, `#level`, `#nesting_level` , `#set_level`
   - this allows a more unified handling with `Expression::Base` instances